@@ -1,97 +1,103 @@
-import 'package:dart_rss/dart_rss.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:url_launcher/url_launcher.dart' as url;
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:snow_weather_info/core/widgets/app_web_page.dart';
-import 'package:snow_weather_info/data/data_notifier.dart';
+import 'package:snow_weather_info/data/sources/avalanche_api.dart';
 import 'package:snow_weather_info/extensions/atom_item.dart';
+import 'package:url_launcher/url_launcher.dart' as url;
 
-class AvalancheListWidget extends StatelessWidget {
+class AvalancheListWidget extends ConsumerWidget {
   const AvalancheListWidget({
     super.key,
   });
 
   @override
-  Widget build(BuildContext context) {
-    final feedItems = context.select<DataNotifier, List<AtomItem>>(
-      (n) => n.avalancheFeed?.items ?? [],
+  Widget build(BuildContext context, WidgetRef ref) {
+    final feedItems = ref.watch(
+      apiAvalancheProvider,
     );
     final titleStyle = TextStyle(
-      color: Theme.of(context).textTheme.bodyText2?.color,
+      color: Theme.of(context).textTheme.bodyMedium?.color,
       fontStyle: FontStyle.normal,
     );
 
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        if (feedItems.isNotEmpty)
-          Expanded(
-            child: ListView.builder(
-              itemCount: feedItems.length,
-              itemBuilder: (context, index) {
-                final item = feedItems[index];
+    return feedItems.when(
+      data: (data) => Column(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          if (data != null && data.items.isNotEmpty)
+            Expanded(
+              child: ListView.builder(
+                itemCount: data.items.length,
+                itemBuilder: (context, index) {
+                  final item = data.items[index];
 
-                return Card(
-                  margin: const EdgeInsets.all(
-                    8,
-                  ),
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                  elevation: 5,
-                  child: ListTile(
-                    title: Text(
-                      item.shortTitle,
-                      style: titleStyle,
+                  return Card(
+                    margin: const EdgeInsets.all(
+                      8,
                     ),
-                    subtitle: Text(
-                      '${item.date} : ${item.massif}',
-                      style: titleStyle,
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(20),
                     ),
-                    trailing: item.url.isNotEmpty
-                        ? const Icon(Icons.navigate_next)
-                        : const SizedBox(),
-                    onTap: () => Navigator.push(
-                      context,
-                      MaterialPageRoute<Widget>(
-                        builder: (context) => AppWebPage(
-                          title: item.shortTitle,
-                          url: item.url,
-                          canIsOpen: true,
+                    elevation: 5,
+                    child: ListTile(
+                      title: Text(
+                        item.shortTitle,
+                        style: titleStyle,
+                      ),
+                      subtitle: Text(
+                        '${item.date} : ${item.massif}',
+                        style: titleStyle,
+                      ),
+                      trailing: item.url.isNotEmpty
+                          ? const Icon(Icons.navigate_next)
+                          : const SizedBox(),
+                      onTap: () => Navigator.push(
+                        context,
+                        MaterialPageRoute<Widget>(
+                          builder: (context) => AppWebPage(
+                            title: item.shortTitle,
+                            url: item.url,
+                            canIsOpen: true,
+                          ),
                         ),
                       ),
                     ),
-                  ),
-                );
-              },
+                  );
+                },
+              ),
+            )
+          else
+            const Padding(
+              padding: EdgeInsets.all(20),
+              child: Text("Pas de donnée d'avalanche récente"),
             ),
-          )
-        else
-          const Padding(
-            padding: EdgeInsets.all(20),
-            child: Text("Pas de donnée d'avalanche récente"),
-          ),
-        const _AttributionWidget(),
-      ],
+          const _AttributionWidget(),
+        ],
+      ),
+      loading: () => const Center(
+        child: CircularProgressIndicator(),
+      ),
+      error: (_, __) => const Padding(
+        padding: EdgeInsets.all(20),
+        child: Text("Erreur lors du chargement des données"),
+      ),
     );
   }
 }
 
 class _AttributionWidget extends StatelessWidget {
-  const _AttributionWidget({
-    super.key,
-  });
+  const _AttributionWidget();
 
   @override
   Widget build(BuildContext context) {
-    final bodyText2Color = Theme.of(context).textTheme.bodyText2?.color;
-    final bodyText1Color = Theme.of(context).textTheme.bodyText1?.color;
+    final bodyText2Color = Theme.of(context).textTheme.bodyMedium?.color;
+    final bodyText1Color = Theme.of(context).textTheme.bodyLarge?.color;
 
     return ColoredBox(
-      color: Theme.of(context).backgroundColor,
+      color: Theme.of(context).colorScheme.background,
       child: Padding(
-        padding: EdgeInsets.all(10),
+        padding: const EdgeInsets.all(10),
         child: RichText(
           text: TextSpan(
             children: [
