@@ -9,85 +9,85 @@ abstract class ConstantDatalist {
     //alpes nord
     Nivose(
       'Aiguilles rouges',
-      LatLng(45.986125, 6.895937),
+      const LatLng(45.986125, 6.895937),
       2330,
       'AIGRG',
     ),
     Nivose(
       'Bellecote',
-      LatLng(45.48881, 6.77163),
+      const LatLng(45.48881, 6.77163),
       3000,
       'BELLE',
     ),
     Nivose(
       'Le Chevril',
-      LatLng(45.50179, 6.95604),
+      const LatLng(45.50179, 6.95604),
       2560,
       'CHEVR',
     ),
     Nivose(
       'Bonneval',
-      LatLng(45.35175, 7.05132),
+      const LatLng(45.35175, 7.05132),
       2720,
       'BONNE',
     ),
     Nivose(
       'Les rochilles',
-      LatLng(45.08467, 6.47127),
+      const LatLng(45.08467, 6.47127),
       2450,
       'ROCHI',
     ),
     Nivose(
       'Allant',
-      LatLng(45.66345, 6.21195),
+      const LatLng(45.66345, 6.21195),
       1630,
       'ALLAN',
     ),
     Nivose(
       'Grande Parei',
-      LatLng(45.61463, 6.64750),
+      const LatLng(45.61463, 6.64750),
       2240,
       'PAREI',
     ),
     Nivose(
       'Col de Porte',
-      LatLng(45.2952742, 5.7651929),
+      const LatLng(45.2952742, 5.7651929),
       1325,
       'COLPO',
     ),
     Nivose(
       'St Hilaire',
-      LatLng(45.31449, 5.86462),
+      const LatLng(45.31449, 5.86462),
       1700,
       'STHIL',
     ),
     Nivose(
       'Aigleton',
-      LatLng(45.2311598, 6.0381140),
+      const LatLng(45.2311598, 6.0381140),
       2240,
       'AIGLE',
     ),
     Nivose(
       'Le Gua',
-      LatLng(45.01517, 5.58951),
+      const LatLng(45.01517, 5.58951),
       1600,
       'LEGUA',
     ),
     Nivose(
       'Les Ecrins',
-      LatLng(44.93681, 6.34602),
+      const LatLng(44.93681, 6.34602),
       2940,
       'ECRIN',
     ),
     Nivose(
       'La Meije',
-      LatLng(45.01255, 6.26485),
+      const LatLng(45.01255, 6.26485),
       3100,
       'MEIJE',
     ),
     Nivose(
       'Galibier',
-      LatLng(45.05733, 6.37739),
+      const LatLng(45.05733, 6.37739),
       2559,
       'GALIB',
     )
@@ -95,31 +95,31 @@ abstract class ConstantDatalist {
     ,
     Nivose(
       'Orcières',
-      LatLng(44.71857, 6.33378),
+      const LatLng(44.71857, 6.33378),
       2294,
       'ORCIE',
     ),
     Nivose(
       'Col Agnel',
-      LatLng(44.68873, 6.97625),
+      const LatLng(44.68873, 6.97625),
       2630,
       'AGNEL',
     ),
     Nivose(
       'Restefond',
-      LatLng(44.34205, 6.79871),
+      const LatLng(44.34205, 6.79871),
       2550,
       'RESTE',
     ),
     Nivose(
       'Millefonts',
-      LatLng(44.11819, 7.19226),
+      const LatLng(44.11819, 7.19226),
       2430,
       'MILLE',
     ),
     Nivose(
       'Parpaillon',
-      LatLng(44.49522, 6.63650),
+      const LatLng(44.49522, 6.63650),
       2545,
       'PARPA',
     )
@@ -127,13 +127,13 @@ abstract class ConstantDatalist {
     ,
     Nivose(
       'Sponde',
-      LatLng(42.32448, 8.88478),
+      const LatLng(42.32448, 8.88478),
       1980,
       'SPOND',
     ),
     Nivose(
       'Maniccia',
-      LatLng(42.20837, 9.04888),
+      const LatLng(42.20837, 9.04888),
       2360,
       'MANIC',
     )
@@ -141,49 +141,49 @@ abstract class ConstantDatalist {
     ,
     Nivose(
       'Maupas',
-      LatLng(42.71428, 0.54995),
+      const LatLng(42.71428, 0.54995),
       2430,
       'MAUPA',
     ),
     Nivose(
       'Port d`Aula',
-      LatLng(42.76993, 1.11493),
+      const LatLng(42.76993, 1.11493),
       2140,
       'PAULA',
     ),
     Nivose(
       'Canigou',
-      LatLng(42.53322, 2.46096),
+      const LatLng(42.53322, 2.46096),
       2150,
       'CANIG',
     ),
     Nivose(
       'Hospitalet',
-      LatLng(42.57778, 1.81140),
+      const LatLng(42.57778, 1.81140),
       2293,
       'HOSPI',
     ),
     Nivose(
       'Puigmal',
-      LatLng(42.38318, 2.09143),
+      const LatLng(42.38318, 2.09143),
       2467,
       'PUIGN',
     ),
     Nivose(
       'Soum Couy',
-      LatLng(42.96157, -0.71995),
+      const LatLng(42.96157, -0.71995),
       2150,
       'SOUMC',
     ),
     Nivose(
       'Lac d’ardiden',
-      LatLng(42.84828, -0.06471),
+      const LatLng(42.84828, -0.06471),
       2445,
       'LARDI',
     ),
     Nivose(
       'Aiguillettes',
-      LatLng(42.74692, 0.18128),
+      const LatLng(42.74692, 0.18128),
       2120,
       'AIGTE',
     ),
