@@ -1,9 +1,15 @@
-extension DoubleExtension on double {
-  String toStringSnowHeigth() {
-    return '${(this * 100).toStringAsFixed(1)}cm';
-  }
+import 'package:intl/intl.dart';
 
-  String toStringTemperature() {
-    return '${toStringAsFixed(1)}°C';
-  }
+extension DoubleExtension on double {
+  String get toStringSnowHeigth => '${toCm.toStringAsFixed(1)}cm';
+
+  String get toStringTemperature => '${toStringAsFixed(1)}°C';
+
+  double get toCm => this * 100;
+
+  String get toDate => DateFormat.Md().format(
+        DateTime.fromMillisecondsSinceEpoch(
+          toInt(),
+        ),
+      );
 }
