@@ -1,202 +1,270 @@
-import 'package:charts_flutter/flutter.dart' as charts;
+import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:snow_weather_info/extensions/double.dart';
 import 'package:snow_weather_info/model/data_station.dart';
+import 'package:snow_weather_info/modules/data_station/view.dart';
 
-class DataStationChart extends StatelessWidget {
+final _snowColor = Colors.blue.shade600;
+final _snowHeightColor = Colors.blue.shade300.withOpacity(0.5);
+const _tempColor = Colors.red;
+
+class DataStationChart extends ConsumerWidget {
   const DataStationChart({
     super.key,
   });
+  static final _xOffset = const Duration(hours: 12).inMilliseconds;
 
-  @override
-  Widget build(BuildContext context) {
-    final List<DataStation> data = context.watch<List<DataStation>>();
+  LineChartData _getTempData(List<DataStation> datas) {
+    final tempData = <FlSpot>[];
+    final minTempData = <FlSpot>[];
+    final maxTempData = <FlSpot>[];
+    final min = datas.firstOrNull?.date.millisecondsSinceEpoch.toDouble() ?? 0;
+    final max = datas.lastOrNull?.date.millisecondsSinceEpoch.toDouble() ?? 0;
 
-    final List<TimeSeriesData> tsdatasnow = [];
-    final List<TimeSeriesData> tsdatanewsnow = [];
-    final List<TimeSeriesData> tsdatasnowtemperature = [];
-    final List<TimeSeriesDataTemp> tsdatatemperature = [];
-
-    // Axis Style
-    final brightness = Theme.of(context).brightness;
-    final charts.Color textColor;
-    final lineColor = charts.MaterialPalette.gray.shade400;
-    textColor = brightness == Brightness.dark
-        ? charts.MaterialPalette.white
-        : charts.MaterialPalette.black;
-    final primaryAxis = charts.NumericAxisSpec(
-      showAxisLine: false,
-      renderSpec: charts.GridlineRendererSpec(
-        lineStyle: charts.LineStyleSpec(color: lineColor),
-        labelStyle: charts.TextStyleSpec(
-          color: textColor,
+    for (var data in datas) {
+      tempData.add(
+        FlSpot(
+          data.date.millisecondsSinceEpoch.toDouble(),
+          data.temperature ?? 0,
         ),
-      ),
-    );
-    final domainAxis = charts.DateTimeAxisSpec(
-      renderSpec: charts.SmallTickRendererSpec(
-        labelStyle: charts.TextStyleSpec(
-          color: textColor,
-        ),
-      ),
-      tickFormatterSpec: const charts.AutoDateTimeTickFormatterSpec(
-        day: charts.TimeFormatterSpec(
-          format: 'd',
-          transitionFormat: 'dd/MM',
-        ),
-      ),
-    );
+      );
 
-    if (data.isNotEmpty) {
-      for (final d in data) {
-        if (d.snowHeight != null) {
-          tsdatasnow.add(TimeSeriesData(d.date, d.snowHeight! * 100));
-        }
-        if (d.temperature != null) {
-          tsdatatemperature.add(TimeSeriesDataTemp(
-            d.date,
-            d.temperature!,
-            d.temperatureMin24 != null ? d.temperatureMin24! : d.temperature!,
-            d.temperatureMax24 != null ? d.temperatureMax24! : d.temperature!,
-          ));
-        }
-        if (d.temperatureSnow != null) {
-          tsdatasnowtemperature.add(TimeSeriesData(d.date, d.temperatureSnow!));
-        }
-        if (d.snowNewHeight != null) {
-          tsdatanewsnow.add(TimeSeriesData(d.date, d.snowNewHeight! * 100));
-        }
+      final minTemp = data.temperatureMin24;
+      if (minTemp != null) {
+        minTempData.add(
+          FlSpot(
+            data.date.millisecondsSinceEpoch.toDouble(),
+            minTemp,
+          ),
+        );
       }
-    } else {
-      // Dummy list to prevent data = NULL
-      tsdatasnow.add(TimeSeriesData(DateTime.now(), 0));
+      final maxTemp = data.temperatureMax24;
+      if (maxTemp != null) {
+        maxTempData.add(
+          FlSpot(
+            data.date.millisecondsSinceEpoch.toDouble(),
+            maxTemp,
+          ),
+        );
+      }
     }
 
-    final seriesSnow = [
-      charts.Series<TimeSeriesData, DateTime>(
-        id: 'SnowHeight',
-        displayName: 'Neige',
-        colorFn: (_, __) => charts.MaterialPalette.blue.shadeDefault,
-        domainFn: (TimeSeriesData data, _) => data.time,
-        measureFn: (TimeSeriesData data, _) => data.data,
-        data: tsdatasnow,
-      ),
-      charts.Series<TimeSeriesData, DateTime>(
-        id: 'SnowNewHeight',
-        displayName: 'Neige fraîche',
-        colorFn: (_, __) => charts.MaterialPalette.blue.shadeDefault,
-        domainFn: (TimeSeriesData data, _) => data.time,
-        measureFn: (TimeSeriesData data, _) => data.data,
-        data: tsdatanewsnow,
-      )..setAttribute(charts.rendererIdKey, 'customBar'),
-    ];
+    final temperatureLineData = LineChartBarData(
+      spots: tempData,
+      isCurved: true,
+      color: _tempColor,
+    );
+    final temperatureMaxLineData = LineChartBarData(
+      spots: minTempData,
+      isCurved: true,
+      show: false,
+      dotData: const FlDotData(show: false),
+    );
+    final temperatureMinLineData = LineChartBarData(
+      spots: maxTempData,
+      isCurved: true,
+      show: false,
+      dotData: const FlDotData(show: false),
+    );
 
-    final seriesTemperature = [
-      charts.Series<TimeSeriesDataTemp, DateTime>(
-        id: 'Temperature',
-        displayName: 'Température',
-        colorFn: (_, __) => charts.MaterialPalette.red.shadeDefault,
-        domainFn: (TimeSeriesDataTemp data, _) => data.time,
-        measureFn: (TimeSeriesDataTemp data, _) => data.temp,
-        measureLowerBoundFn: (TimeSeriesDataTemp data, _) => data.tempMin,
-        measureUpperBoundFn: (TimeSeriesDataTemp data, _) => data.tempMax,
-        data: tsdatatemperature,
+    return LineChartData(
+      lineBarsData: [
+        temperatureLineData,
+        temperatureMaxLineData,
+        temperatureMinLineData,
+      ],
+      minX: min - _xOffset,
+      maxX: max + _xOffset,
+      betweenBarsData: [
+        BetweenBarsData(
+          fromIndex: 0,
+          toIndex: 1,
+          color: _tempColor.withOpacity(0.5),
+        ),
+        BetweenBarsData(
+          fromIndex: 0,
+          toIndex: 2,
+          color: _tempColor.withOpacity(0.5),
+        )
+      ],
+      titlesData: FlTitlesData(
+        bottomTitles: _bottomTitles(temperatureLineData),
+        topTitles: _noTitlesWidget(),
+        rightTitles: _noTitlesWidget(),
       ),
-      charts.Series<TimeSeriesData, DateTime>(
-        id: 'Temperature Snow',
-        displayName: 'Température Neige',
-        colorFn: (_, __) => charts.MaterialPalette.pink.shadeDefault,
-        domainFn: (TimeSeriesData data, _) => data.time,
-        measureFn: (TimeSeriesData data, _) => data.data,
-        data: tsdatasnowtemperature,
-      ),
-    ];
+    );
+  }
 
-    return ListView(
-      primary: false,
-      shrinkWrap: true,
+  LineChartData _getSnowData(
+    List<DataStation> datas,
+  ) {
+    final tempData = <FlSpot>[];
+    final List<List<FlSpot>> tempHeightData = [];
+    final min = datas.firstOrNull?.date.millisecondsSinceEpoch.toDouble() ?? 0;
+    final max = datas.lastOrNull?.date.millisecondsSinceEpoch.toDouble() ?? 0;
+
+    for (final data in datas) {
+      final x = data.date.millisecondsSinceEpoch.toDouble();
+      tempData.add(
+        FlSpot(
+          x,
+          data.snowHeight?.toCm ?? 0,
+        ),
+      );
+
+      final dataHeight = data.snowNewHeight?.toCm ?? 0;
+      if (dataHeight > 0) {
+        tempHeightData.add([
+          FlSpot(
+            x,
+            dataHeight,
+          ),
+          FlSpot(
+            x,
+            0,
+          ),
+        ]);
+      }
+    }
+
+    final snowLineData = LineChartBarData(
+      spots: tempData,
+      isCurved: true,
+      color: _snowColor,
+    );
+
+    final snowHeightLineData = tempHeightData.map(
+      (d) => LineChartBarData(
+        spots: d,
+        barWidth: 20,
+        dotData: const FlDotData(show: false),
+        color: _snowHeightColor,
+      ),
+    );
+
+    final maxY = snowLineData.mostTopSpot.y + 10;
+
+    return LineChartData(
+      lineBarsData: [
+        snowLineData,
+        ...snowHeightLineData,
+      ],
+      minX: min - _xOffset,
+      maxX: max + _xOffset,
+      minY: 0,
+      maxY: maxY,
+      titlesData: FlTitlesData(
+        bottomTitles: _bottomTitles(snowLineData),
+        topTitles: _noTitlesWidget(),
+        rightTitles: _noTitlesWidget(),
+      ),
+      borderData: FlBorderData(
+        border: const Border(
+          top: BorderSide.none,
+          right: BorderSide.none,
+          left: BorderSide(width: 1),
+          bottom: BorderSide(width: 1),
+        ),
+      ),
+    );
+  }
+
+  AxisTitles _bottomTitles(LineChartBarData data) {
+    late double? interval;
+    if (data.spots.length > 1) {
+      interval = data.mostRightSpot.x - data.mostLeftSpot.x;
+    } else {
+      interval = data.spots.firstOrNull?.x;
+    }
+    return AxisTitles(
+      sideTitles: SideTitles(
+        interval: interval,
+        showTitles: true,
+        getTitlesWidget: (value, meta) {
+          return Text(value.toDate);
+        },
+      ),
+    );
+  }
+
+  AxisTitles _noTitlesWidget() {
+    return const AxisTitles(sideTitles: SideTitles());
+  }
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final datas = ref.watch(currentDataStationProvider);
+    const graphPadding = EdgeInsets.only(right: 40, top: 40);
+    final snowDatas = _getSnowData(datas);
+
+    return Column(
       children: [
         Container(
-          constraints: const BoxConstraints(
-            maxHeight: 300,
-          ),
-          child: charts.TimeSeriesChart(
-            seriesSnow,
-            animate: true,
-            animationDuration: const Duration(milliseconds: 800),
-            behaviors: [
-              charts.SeriesLegend(
-                position: charts.BehaviorPosition.bottom,
-                showMeasures: true,
-                horizontalFirst: false,
-                measureFormatter: (value) {
-                  return value == null ? '-' : '${value.toStringAsFixed(1)}cm';
-                },
-              ),
-            ],
-            layoutConfig: charts.LayoutConfig(
-              leftMarginSpec: charts.MarginSpec.fixedPixel(30),
-              topMarginSpec: charts.MarginSpec.fixedPixel(10),
-              rightMarginSpec: charts.MarginSpec.fixedPixel(10),
-              bottomMarginSpec: charts.MarginSpec.fixedPixel(10),
-            ),
-            primaryMeasureAxis: primaryAxis,
-            domainAxis: domainAxis,
-            customSeriesRenderers: [
-              charts.BarRendererConfig(
-                // ID used to link series to this renderer.
-                customRendererId: 'customBar',
-                cornerStrategy: const charts.ConstCornerStrategy(30),
-                fillPattern: charts.FillPatternType.forwardHatch,
-              ),
-            ],
-          ),
+          height: 300,
+          padding: graphPadding,
+          child: LineChart(snowDatas),
+        ),
+        const SizedBox(
+          height: 10,
+        ),
+        _LegendChart(
+          color: _snowColor,
+          label: 'Neige',
+        ),
+        _LegendChart(
+          color: _snowHeightColor,
+          label: 'Neige fraîche',
+        ),
+        const SizedBox(
+          height: 10,
         ),
         Container(
-          constraints: const BoxConstraints(
-            maxHeight: 300,
-          ),
-          child: charts.TimeSeriesChart(
-            seriesTemperature,
-            animate: true,
-            animationDuration: const Duration(milliseconds: 800),
-            behaviors: [
-              charts.SeriesLegend(
-                position: charts.BehaviorPosition.bottom,
-                showMeasures: true,
-                horizontalFirst: false,
-                measureFormatter: (value) {
-                  return value == null ? '-' : '${value.toStringAsFixed(1)}°C';
-                },
-              ),
-            ],
-            layoutConfig: charts.LayoutConfig(
-              leftMarginSpec: charts.MarginSpec.fixedPixel(30),
-              topMarginSpec: charts.MarginSpec.fixedPixel(10),
-              rightMarginSpec: charts.MarginSpec.fixedPixel(10),
-              bottomMarginSpec: charts.MarginSpec.fixedPixel(10),
-            ),
-            primaryMeasureAxis: primaryAxis,
-            domainAxis: domainAxis,
-          ),
+          height: 300,
+          padding: graphPadding,
+          child: LineChart(_getTempData(datas)),
+        ),
+        const SizedBox(
+          height: 10,
+        ),
+        const _LegendChart(
+          color: _tempColor,
+          label: 'Température',
+        ),
+        const SizedBox(
+          height: 10,
         ),
       ],
     );
   }
 }
 
-class TimeSeriesData {
-  TimeSeriesData(this.time, this.data);
+class _LegendChart extends StatelessWidget {
+  const _LegendChart({
+    required this.color,
+    required this.label,
+  });
 
-  final DateTime time;
-  final double data;
-}
+  final String label;
+  final Color color;
 
-class TimeSeriesDataTemp {
-  TimeSeriesDataTemp(this.time, this.temp, this.tempMin, this.tempMax);
-
-  final DateTime time;
-  final double temp;
-  final double tempMin;
-  final double tempMax;
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 20),
+      child: Row(
+        children: [
+          SizedBox(
+            width: 12,
+            height: 12,
+            child: ColoredBox(color: color),
+          ),
+          const SizedBox(
+            width: 8,
+          ),
+          Text(label),
+        ],
+      ),
+    );
+  }
 }
