@@ -1,32 +1,54 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter/scheduler.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:share/share.dart';
-import 'package:snow_weather_info/data/data_notifier.dart';
 import 'package:snow_weather_info/model/station.dart';
+import 'package:snow_weather_info/modules/map/map_widget.dart';
+import 'package:snow_weather_info/provider/favorite_station.dart';
 
-class NivosePage extends StatefulWidget {
+part 'nivose_page.g.dart';
+
+@riverpod
+bool _isFavorite(Ref ref, String codeMF) {
+  final stations = ref.watch(favoriteStationProvider);
+
+  return stations.any(
+    (element) => switch (element) {
+      Nivose nivose => nivose.codeMF == codeMF,
+      _ => false,
+    },
+  );
+}
+
+class NivosePage extends ConsumerStatefulWidget {
   const NivosePage({
     super.key,
     required this.nivose,
   });
-
   final Nivose nivose;
 
   @override
-  _NivosePageState createState() => _NivosePageState();
+  ConsumerState<ConsumerStatefulWidget> createState() => _NivosePageState();
 }
 
-class _NivosePageState extends State<NivosePage> {
+class _NivosePageState extends ConsumerState<NivosePage> {
   @override
   void initState() {
     super.initState();
-    context.read<DataNotifier>().currentMapLoc = widget.nivose.position;
+    SchedulerBinding.instance.addPostFrameCallback((_) {
+      if (mounted) {
+        ref.read(currentMapLocProvider.notifier).setLocation(
+              widget.nivose.position,
+            );
+      }
+    });
   }
 
   @override
   Widget build(BuildContext context) {
-    final isFavorite = context.select(
-      (DataNotifier n) => n.isFavorite(widget.nivose),
+    final isFavorite = ref.watch(
+      _isFavoriteProvider(widget.nivose.codeMF),
     );
 
     return Scaffold(
@@ -36,8 +58,8 @@ class _NivosePageState extends State<NivosePage> {
           IconButton(
             icon: Icon(isFavorite ? Icons.favorite : Icons.favorite_border),
             onPressed: () => setState(
-              () => context
-                  .read<DataNotifier>()
+              () => ref
+                  .read(favoriteStationProvider.notifier)
                   .addOrRemoveFavoriteStation(widget.nivose),
             ),
           ),
@@ -73,7 +95,6 @@ class _NivosePageState extends State<NivosePage> {
 
 class _NetworkImage extends StatelessWidget {
   const _NetworkImage({
-    super.key,
     required this.url,
   });
 
